openrisc: dts: Split simple smp dts to dts and dtsi
Split out the common memory, CPU and PIC definitions of the simple SMP
system to a DTSI file which we will later use for our De0 Nano multicore
board device tree.  We also take this opportunity to swich underscores
to dashes as that seems to be the more common convention for DTS files.

Signed-off-by: Stafford Horne <shorne@gmail.com>

[ upstream commit: 11659e4c3a1463ba8e49078c009228d7e1f54955 ]