dt-bindings: mfd: Add Realtek RTD1xxx system controllers
The Realtek system controllers are already in use with only generic
'"syscon", "simple-mfd"' compatibles and are missing specific
compatibles as required. Add a schema with specific compatibles. The
labels used in .dts files serve as the basis for the names.

It's doubtful the child nodes upstream are complete and I don't have
documentation, so the specific child nodes aren't documented here. Of
the ones in use, bindings already exist for them.

Link: https://patch.msgid.link/20251215212624.3319681-1-robh@kernel.org
Signed-off-by: Rob Herring (Arm) <robh@kernel.org>

[ upstream commit: c0238bbba7c5a45afbfe61a31c3ba322f744feb0 ]